Refactor database configuration in autoscaler API and metrics forwarder
 • Centralize database configuration logic into db.ConfigureDb and db.ConfigureStoredProcedureDb
 • Remove individual configurePolicyDb and configureBindingDb functions
 • Move DatabaseConfig struct to a new models.go file in the db package
 • Update calls to database configuration in both API and metrics forwarder to use the new centralized functions
